Scripting lab 7
I had some problems and messaged you and you said just to do the challenge, but the challenge is to do the same thing for another value, like health or mana, so I figured out my own way to do the health bar. I know its not very scalable, so consider it a temporary solution.